+++ ament_cmake_test/CMakeLists.txt
@@ -3,6 +3,9 @@ cmake_minimum_required(VERSION 3.5)
 project(ament_cmake_test NONE)
 
 find_package(ament_cmake_core REQUIRED)
+find_package(ament_cmake_python REQUIRED)
+
+ament_python_install_package(${PROJECT_NAME})
 
 ament_package(
   CONFIG_EXTRAS "ament_cmake_test-extras.cmake"
