+++ ament_cmake_test/package.xml
@@ -8,6 +8,7 @@
   <license>Apache License 2.0</license>
 
   <buildtool_depend>ament_cmake_core</buildtool_depend>
+  <buildtool_depend>ament_cmake_python</buildtool_depend>
 
   <buildtool_export_depend>ament_cmake_core</buildtool_export_depend>
